修改I的数据类型转换规则：向下取整; 修改程序中初始光子数数据类型

+++ Source/DataType.h
@@ -41,7 +41,7 @@ struct CTScanParas
 
 	float dSize;							 // 探测器分辨率   mm
 
-	uint32 I0Val;							// 亮场
+	float I0Val;							// 亮场
 	int spectrumStep;						// 能谱间隔
 	int specEnergyNum;						// 离散化能谱个数
